+++ Utils/CHANGELOG.md
@@ -4,10 +4,19 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- CPMath (general math functions)
+- ProjectileMath (math functions related to projectiles)
+
+### Changed
+### Removed
+
 ## [0.0.1] - 2021-01-24
 ### Added
 - FiniteStateMachine class
 - ObjectPool class and tests
 - Extensions class with a collection of extension methods
 
+[Unreleased]: https://github.com/CheesePie13/UnityPackages/compare/v0.0.1...main
 [0.0.1]: https://github.com/CheesePie13/UnityPackages/releases/tag/utils-v0.0.1